Fix lint

+++ lib/model/presentation/int_representation.dart
@@ -3,7 +3,7 @@ import 'package:intl/intl.dart';
 class IntRepresentation {
   static String toReadableValue(String value) {
     try {
-      return NumberFormat.compact().format(int.parse(value));
+      return NumberFormat.compact().format(double.parse(value));
     } catch (error) {
       return value;
     }
